Add cast to please debug build
    XUtils.c:292:21: error: comparison of integers of different signs: 'ssize_t' (aka 'long') and 'size_t' (aka 'unsigned long') [-Werror,-Wsign-compare]
      292 |          assert(res <= count);
          |                 ~~~ ^  ~~~~~

+++ XUtils.c
@@ -289,7 +289,7 @@ static ssize_t readfd_internal(int fd, void* buffer, size_t count) {
       }
 
       if (res > 0) {
-         assert(res <= count);
+         assert((size_t)res <= count);
 
          buffer = ((char*)buffer) + res;
          count -= (size_t)res;
